BTC _ RNN algorithm
Bitcoin (in English: Bitcoin) (abbreviated: BTC; symbol: ₿) is a decentralized digital currency, without a central bank or single management, which can be sent peer-to-peer through the Bitcoin network without the need for an intermediary. ] Transactions are cryptographically verified by a network and recorded in a public distributed ledger called a blockchain. This cryptocurrency was invented in 2008 by an unknown person or group named Satoshi Nakamoto.[5] The use of this currency started in 2009[6] after the release of its open source software.

In the process of cryptocurrency mining, the mining person (miner) is rewarded with Bitcoin, which can be used to buy other currencies, products and services. Bitcoin has been criticized for being used in illegal transactions, consuming a lot of electricity in mining (and thus the carbon footprint), price volatility, and theft from exchanges. In the past, some investors and economists have described Bitcoin as a speculative bubble, but some have used it as an investment. Several regulatory agencies have issued warnings to investors about Bitcoin.[7][8][9] In September 2021, El Salvador officially accepted Bitcoin as legal tender, becoming the first country to do so.[10] ]

The word bitcoin was defined in a white paper published on October 31, 2008. [3][11] It is a combination of bit and coin. [12] There is no single standard for writing the word bitcoin. Some sources use Bitcoin (Bitcoin) with uppercase letters to refer to technology and computer network and Bitcoin (bitcoin) with lowercase letters for the unit of account.[13] The Wall Street Journal,[14] and Oxford English Dictionary[12]. They support writing bitcoin in lowercase letters in all cases.